Use `Path::to_string_lossy`
The existing code uses a conversion that signals errors when converting from the platform specific file-system encoding to unicode. It then panics on a failure. The error message doesn't match the actual error. This commit simplifies and corrects this by using to_string_lossy which will do the same conversion, but replaces illegal characters with ? - a well known behaviour from other languages.