Move transformation from Heroku to Caddy configuration into a TryFrom
Now that we separated TOML parsing, Heroku config and Caddy config, we can create a higher-level transformation between the Heroku and Caddy configuration types. This is usually done with the TryFrom trait and we do exactly this here. This separates the concerns even further as the tests for the TryFrom can now work with static HerokuWebServerConfig values.